docs(app): add copy markdown button (#4369)
